// front-coin-wallet/src/app/services/com.b2beyond.api.user/model/user.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { UserDetails } from './userDetails';


export interface User {
    websiteId?: number;
    userId?: string;
    password?: string;
    secretKey?: string;
    active?: boolean;
    passwordResetToken?: string;
    activationToken?: string;
    ip?: string;
    authenticationToken?: string;
    authenticationRefreshToken?: string;
    userDetails?: UserDetails;
    enabled2FA?: boolean;
    code2FA?: number;
    valid?: boolean;
}